// src/app/Models/Propiedades.ts
export interface Propiedad{
  IDPropiedad: number;
	TerrenoM2: number;
	Descripcion: string;
	CantCuartos: number;
	CantBanhos: number;
	CantDormitorios: number;
	NotasExtra: string;
	Precio: number;
	FechaDeCompra: string;
	IDDireccionPropiedadesP: number;
	IDEstatusPropiedadP: number;
	IDPropietarioP: number;
}